Added video_genre_cleanup fixup
The video_genre_cleanup fixup walks through the movie and show items of the
configured library sections, and cleans up the "Genres" field of each item.

It can consolidate genres into a defined list of genres, remove useless
genres, and set a default genre if the list of genres is empty.

These changes can be configured in the config file.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>